[REFACTOR] 파일명 변경  및 불필요한 값 삭제
related to : #135
