+++ src/lib/constants/favorites.ts
@@ -1,11 +1,9 @@
 export const FAVORITES_TAB = {
   REGULAR: {
-    key: "regular",
     label: "정기 모임",
     value: "regular",
   },
   QUICK: {
-    key: "quick",
     label: "정기 모임",
     value: "quick",
   },
